pyproject.toml
[tool.poetry]
name = "reliaqueue"
version = "0.1.0"
description = "Имитационное моделирование. Системы массового обслуживания."
authors = ["triks <trik0001000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = {extras = ["all"], version = "^0.111.0"}
django-environ = "^0.11.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
